Improve system shutdown button by display message about shutdown and system working time

+++ src/alert.js
@@ -2,6 +2,7 @@ var gpio = require('onoff').Gpio;
 var lcd = require('../lib/lcd');
 var led = require('../lib/led');
 var log = require('../lib/log.js');
+var startTime = new Date().getTime();
 var buttonLight;
 var buttonOff;
 var buttonStatus;
@@ -12,12 +13,13 @@ exports.launch = function (args, appConfig) {
     config = appConfig;
 
     init();
+    lcd.init();
 
     buttonLight.watch(lcdLight);
     buttonOff.watch(systemOff);
 };
 
-function lcdLight (err, state) {
+function lcdLight(err, state) {
     if(state == 1) {
         if (lcdLightStatus) {
             lcd.lightOff();
@@ -29,7 +31,7 @@ function lcdLight (err, state) {
     }
 }
 
-function init () {
+function init() {
     buttonLight = new gpio(
         config.get('alert_gpio.button_display'),
         'in',
@@ -47,11 +49,28 @@ function init () {
     lcd.init();
 }
 
-function systemOff () {
-    var exec = require('child_process').exec;
+function systemOff(err, state) {
+    if(state == 1) {
+        var uptime = upTime();
+        var exec = require('child_process').exec;
 
-    log.logInfo('System shutdown.');
-    console.log('System is shutting down.');
+        lcd.clear();
+        lcd.displayMessage([
+            'System shutdown',
+            'after: ' + uptime
+        ]);
 
-    exec('sudo shutdown -h now');
+        log.logInfo('System shutdown after: ' + uptime);
+        console.log('System is shutting down.');
+
+        exec('sudo shutdown -h now');
+    }
+}
+
+function upTime() {
+    var currentTime = new Date().getTime();
+    var calc = currentTime - startTime;
+    var diff = new Date(calc);
+
+    return (diff.getHours() -1) + ':' + diff.getMinutes() + ':' + diff.getSeconds();
 }
